+++ CHANGELOG.md
@@ -14,6 +14,7 @@ consistent in the database. This decreases processing latency but does not give 
 - DeltaFile metrics for total/in flight files and bytes
 - DeltaFile metrics graphs added to system summary dashboard
 - Added `/blackhole` endpoint to `deltafi-egress-sink`.  The endpoint will always return a 200 and never write content to disk, acting as a noop egress destination.  The endpoint will take a latency parameter to add latency to the post response (i.e. `/blackhole?latency=0.1` to add 100ms latency)
+- MergeContentJoinAction that merges content by binary concatenation, TAR, ZIP, AR, TAR.GZ, or TAR.XZ.
 - Added ingress, survey, and error documentation
 
 ### Changed